requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==4.2.10
django-bootstrap5==23.4
django-cpf-cnpj==1.0.0
djangorestframework==3.14.0
fonttools==4.49.0
idna==3.6
importlib_resources==6.1.2
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pyparsing==3.1.1
python-dateutil==2.9.0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
zipp==3.17.0
